app/requirements-dev.txt
flake8==7.1.0
pathspec==0.12.1
black==24.4.2
pycodestyle==2.12.0
mypy-extensions==1.0.0
mccabe==0.7.0
pyflakes==3.2.0
click==8.1.7
isort==5.13.2
platformdirs==4.2.2
packaging==24.1
pluggy==1.5.0
iniconfig==2.0.0
attrs==23.2.0
pytest==7.2.2
pytest-django==4.5.2
